Fully functional website with user authentification and unique workouts for users

+++ front-end/src/components/WorkoutDetails.jsx
@@ -1,19 +1,30 @@
 import { useWorkoutContext } from '../hooks/useWorkoutContext';
+import { useAuthContext } from '../hooks/useAuthContext';
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
 import {BsTrash} from 'react-icons/bs';
+
 
 const WorkoutDetails = ({workout}) => {
 
     const { dispatch } = useWorkoutContext();
+    const { user } = useAuthContext();
 
     const handleClick = async () => {
+
+        if (!user) {
+            return
+        }
+
         const response = await fetch("/api/workouts/" + workout._id, {
             method: "DELETE",
-        })
-        const json = await response.json();
+            headers: {
+                'Authorization': "Bearer " + user.token,
+        }
+    })
+        const data = await response.json();
 
         if(response.ok) {
-            dispatch({type: 'DELETE_WORKOUT', payload: json});
+            dispatch({type: 'DELETE_WORKOUT', payload: data});
         }
     };
 
@@ -24,8 +35,7 @@ const WorkoutDetails = ({workout}) => {
             <p>Sets: {workout.sets}</p>
             <p>Weight : {workout.weight} kilos</p>
             <p>{formatDistanceToNow(new Date(workout.createdAt), {addSuffix: true})}</p>
-            <button id="delete-btn" onClick={handleClick}><BsTrash/></button>
-            
+            <button className="delete-btn" onClick={handleClick}><BsTrash/></button>
         </div>
     )
 }
